+++ src/main/java/fr/lenny/backend/service/MessageService.java
@@ -4,5 +4,5 @@ import fr.lenny.backend.entity.Message;
 
 public interface MessageService {
     public void addMessage(Message message);
-    public Message getMessage(int messageId);
+    public Message getMessage(Long messageId);
 }
