+++ src/main/java/fr/lenny/backend/service/UserService.java
@@ -3,5 +3,5 @@ package fr.lenny.backend.service;
 import fr.lenny.backend.entity.User;
 
 public interface UserService {
-    public User getUser(int userId);
+    public User getUser(Long userId);
 }
